refactor: 구간 저장 비즈니스를 LineService가 담당하도록 리팩터링
