+++ src/main/java/subway/dto/SectionCreateReader.java
@@ -1,0 +1,9 @@
+package subway.dto;
+
+public interface SectionCreateReader {
+    Long getDownStationId();
+
+    Long getUpStationId();
+
+    Long getDistance();
+}
